requirements.txt
Cython==0.28
google-speech==1.0.16
imutils==0.4.6
numpy==1.14.2
opencv-contrib-python==3.3.0.10
requests==2.18.4
scipy==1.0.0
websocket-client==0.47.0
pygame==1.9.3
pyserial==3.4